requirement.txt
Requirements：

python 3.7
numpy 1.16.5
scikit-learn 0.23.2
pytorch 1.10